+++ projects/angular2-smart-table/src/lib/components/cell/cell-view-mode/view-cell.component.ts
@@ -1,6 +1,7 @@
 import {ChangeDetectionStrategy, Component, Input} from '@angular/core';
 
 import {Cell} from '../../../lib/data-set/cell';
+import {SecurityTrustType} from '../../../pipes/bypass-security-trust.pipe';
 
 @Component({
   selector: 'table-cell-view-mode',
@@ -8,7 +9,7 @@ import {Cell} from '../../../lib/data-set/cell';
   template: `
     <div [ngSwitch]="cell.getColumn().type">
         <custom-view-component *ngSwitchCase="'custom'" [cell]="cell"></custom-view-component>
-        <div *ngSwitchCase="'html'" [innerHTML]="cell.getValue()" [ngClass]="cssClass"></div>
+        <div *ngSwitchCase="'html'" [innerHTML]="cell.getValue() | bypassSecurityTrust: bypassSecurityTrust" [ngClass]="cssClass"></div>
         <div *ngSwitchDefault [ngClass]="cssClass">{{ cell.getValue() }}</div>
     </div>
     `,
@@ -16,6 +17,10 @@ import {Cell} from '../../../lib/data-set/cell';
 export class ViewCellComponent {
 
   @Input() cell!: Cell;
+
+  get bypassSecurityTrust(): SecurityTrustType {
+    return this.cell.getColumn().sanitizer.bypassHtml ? 'html' : 'none';
+  }
 
   get cssClass(): string {
     return this.cell.getColumn().classContent ?? ''