+++ projects/angular2-smart-table/src/lib/components/tbody/cells/expand.component.ts
@@ -1,13 +1,14 @@
 import {ChangeDetectionStrategy, Component, EventEmitter, Input, OnChanges, Output} from "@angular/core";
 import {Row} from "../../../lib/data-set/row";
 import {Grid} from "../../../lib/grid";
+import {SecurityTrustType} from '../../../pipes/bypass-security-trust.pipe';
 
 @Component({
     selector: 'angular2-st-tbody-expand',
     changeDetection: ChangeDetectionStrategy.OnPush,
     template: `
       <a href="#" class="angular2-smart-action angular2-smart-action-expand-expand"
-          [innerHTML]="expandRowButtonContent" (click)="onExpand($event)"></a>
+          [innerHTML]="expandRowButtonContent | bypassSecurityTrust: bypassSecurityTrust" (click)="onExpand($event)"></a>
     `,
   })
   export class TbodyExpandRowComponent implements OnChanges {
@@ -18,6 +19,7 @@ import {Grid} from "../../../lib/grid";
     @Output() onExpandRow = new EventEmitter<any>();
 
     expandRowButtonContent!: string;
+    bypassSecurityTrust: SecurityTrustType = 'none';
 
     constructor(){
     }
@@ -31,5 +33,6 @@ import {Grid} from "../../../lib/grid";
 
     ngOnChanges(){
         this.expandRowButtonContent = this.grid.getSetting('expand.expandRowButtonContent');
+        this.bypassSecurityTrust = this.grid.settings.expand?.sanitizer?.bypassHtml ? 'html' : 'none';
     }
   }