+++ projects/angular2-smart-table/src/lib/components/tbody/cells/save-cancel.component.ts
@@ -3,6 +3,7 @@ import {Component, EventEmitter, Input, OnChanges} from '@angular/core';
 import {Grid} from '../../../lib/grid';
 import {Row} from '../../../lib/data-set/row';
 import {EditCancelEvent, EditConfirmEvent} from '../../../lib/events';
+import {SecurityTrustType} from '../../../pipes/bypass-security-trust.pipe';
 
 @Component({
   // TODO: @breaking-change rename the selector to angular2-st-tbody-save-cancel in the next major version
@@ -23,6 +24,7 @@ export class TbodySaveCancelComponent implements OnChanges {
 
   cancelButtonContent!: string;
   saveButtonContent!: string;
+  bypassSecurityTrust: SecurityTrustType = 'none';
 
   onSave(event: MouseEvent) {
     event.preventDefault();
@@ -46,5 +48,6 @@ export class TbodySaveCancelComponent implements OnChanges {
   ngOnChanges() {
     this.saveButtonContent = this.grid.getSetting('edit.saveButtonContent');
     this.cancelButtonContent = this.grid.getSetting('edit.cancelButtonContent')
+    this.bypassSecurityTrust = this.grid.settings.edit?.sanitizer?.bypassHtml ? 'html' : 'none';
   }
 }