+++ projects/angular2-smart-table/src/lib/components/thead/cells/add-button.component.ts
@@ -3,12 +3,13 @@ import {AfterViewInit, Component, ElementRef, EventEmitter, Input, OnChanges, Ou
 import {Grid} from '../../../lib/grid';
 import {DataSource} from '../../../lib/data-source/data-source';
 import {CreateEvent} from '../../../lib/events';
+import {SecurityTrustType} from '../../../pipes/bypass-security-trust.pipe';
 
 @Component({
   selector: '[angular2-st-add-button]',
   template: `
     <a *ngIf="isActionAdd" href="#" class="angular2-smart-action angular2-smart-action-add-add"
-        [innerHTML]="addNewButtonContent" (click)="onAdd($event)"></a>
+        [innerHTML]="addNewButtonContent | bypassSecurityTrust: bypassSecurityTrust" (click)="onAdd($event)"></a>
   `,
 })
 export class AddButtonComponent implements AfterViewInit, OnChanges {
@@ -19,6 +20,7 @@ export class AddButtonComponent implements AfterViewInit, OnChanges {
 
   isActionAdd!: boolean;
   addNewButtonContent!: string;
+  bypassSecurityTrust: SecurityTrustType = 'none';
 
   constructor(private ref: ElementRef) {
   }
@@ -30,6 +32,7 @@ export class AddButtonComponent implements AfterViewInit, OnChanges {
   ngOnChanges() {
     this.isActionAdd = this.grid.getSetting('actions.add');
     this.addNewButtonContent = this.grid.getSetting('add.addButtonContent');
+    this.bypassSecurityTrust = this.grid.settings.add?.sanitizer?.bypassHtml ? 'html' : 'none';
   }
 
   onAdd(event: any) {