+++ projects/angular2-smart-table/src/lib/components/thead/cells/create-cancel.component.ts
@@ -2,15 +2,16 @@ import {Component, EventEmitter, Input, OnChanges} from '@angular/core';
 
 import {Grid} from '../../../lib/grid';
 import {CreateCancelEvent, CreateConfirmEvent} from '../../../lib/events';
+import {SecurityTrustType} from '../../../pipes/bypass-security-trust.pipe';
 
 @Component({
   // TODO: @breaking-change rename the selector to angular2-st-thead-create-cancel in the next major version
   selector: 'angular2-st-actions',
   template: `
     <a href="#" class="angular2-smart-action angular2-smart-action-add-create"
-        [innerHTML]="createButtonContent" (click)="onCreate($event)"></a>
+        [innerHTML]="createButtonContent | bypassSecurityTrust: bypassSecurityTrust" (click)="onCreate($event)"></a>
     <a href="#" class="angular2-smart-action angular2-smart-action-add-cancel"
-        [innerHTML]="cancelButtonContent" (click)="onCancelCreate($event)"></a>
+        [innerHTML]="cancelButtonContent | bypassSecurityTrust: bypassSecurityTrust" (click)="onCancelCreate($event)"></a>
   `,
 })
 export class TheadCreateCancelComponent implements OnChanges {
@@ -21,6 +22,7 @@ export class TheadCreateCancelComponent implements OnChanges {
 
   createButtonContent!: string;
   cancelButtonContent!: string;
+  bypassSecurityTrust: SecurityTrustType = 'none';
 
   onCreate(event: MouseEvent) {
     event.preventDefault();
@@ -41,5 +43,6 @@ export class TheadCreateCancelComponent implements OnChanges {
   ngOnChanges() {
     this.createButtonContent = this.grid.getSetting('add.createButtonContent');
     this.cancelButtonContent = this.grid.getSetting('add.cancelButtonContent');
+    this.bypassSecurityTrust = this.grid.settings.add?.sanitizer?.bypassHtml ? 'html' : 'none';
   }
 }